Add test for the selector and rename HostRelevantPatches
